Initial extraction of deployments from command_line.py
